% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\docType{class}
\name{DatabaseServer-class}
\alias{DatabaseServer-class}
\alias{DatabaseServer}
\title{DatabaseServer Class}
\description{
A reference class to represent a database server.
}
\section{Fields}{

\describe{
\item{\code{host}}{The host address of the database server.}

\item{\code{port}}{The port number of the database server.}
}}

\section{Methods}{

\describe{
\item{\code{initialize(host = "localhost", port = 3306)}}{Initialize a DatabaseServer object}
}}

