% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{execute_single_query}
\alias{execute_single_query}
\title{Execute a single query on the database}
\usage{
execute_single_query(
  query,
  db_server,
  db_name,
  db_user = NULL,
  db_password = NULL
)
}
\arguments{
\item{query}{The SQL query to be executed.}

\item{db_server}{A DatabaseServer object representing the database server.}

\item{db_name}{The name of the database.}

\item{db_user}{The username for the database connection. Default is NULL.}

\item{db_password}{The password for the database connection. Default is NULL.}
}
\value{
The result of the query.
}
\description{
This function creates a DatabaseConnection object, executes a single query,
and then finalizes the connection.
}
